requirements.txt
Django==1.6.5
South==0.8.4
psycopg2==2.5.2
django-compressor==1.4
django-debug-toolbar==1.0.1
gunicorn==18.0
Fabric==1.8.1
djangorestframework=2.4.0
